feat(connector-wallet-route): goto authorize URL
