+++ connector-wallet-route/WalletProvider.tsx
@@ -1,0 +1,27 @@
+import '@rainbow-me/rainbowkit/styles.css';
+import { getDefaultConfig, RainbowKitProvider } from '@rainbow-me/rainbowkit';
+import { WagmiProvider } from 'wagmi';
+import { mainnet } from 'wagmi/chains';
+import { QueryClientProvider, QueryClient } from '@tanstack/react-query';
+
+const config = getDefaultConfig({
+  appName: 'Apache Answer',
+  projectId: 'xxx',
+  chains: [mainnet],  // There's no on-chain operations, so only ETH mainnet is enough.
+});
+
+const queryClient = new QueryClient();
+
+function WalletProvider({ children }: React.PropsWithChildren) {
+  return (
+    <WagmiProvider config={config}>
+      <QueryClientProvider client={queryClient}>
+        <RainbowKitProvider>
+          {children}
+        </RainbowKitProvider>
+      </QueryClientProvider>
+    </WagmiProvider>
+  );
+}
+
+export default WalletProvider;
