+++ connector-wallet-route/package.json
@@ -42,6 +42,7 @@
   "dependencies": {
     "@rainbow-me/rainbowkit": "^2.2.0",
     "@tanstack/react-query": "^5.59.15",
+    "js-sha256": "0.11.0",
     "viem": "2.x",
     "wagmi": "^2.12.19"
   }
